Move CharWidth into common
