+++ src/Linkifier.ts
@@ -5,7 +5,7 @@
 
 import { ILinkifierEvent, ILinkMatcher, LinkMatcherHandler, ILinkMatcherOptions, ILinkifier, ITerminal, IBufferStringIteratorResult, IMouseZoneManager } from './Types';
 import { MouseZone } from './MouseZoneManager';
-import { getStringCellWidth } from './CharWidth';
+import { getStringCellWidth } from 'common/CharWidth';
 import { EventEmitter2, IEvent } from 'common/EventEmitter2';
 
 /**
